re #438 Region of interest drawing moved to vtkImageVisualizer. Handles to modify ROI still in SegmentationParameterDialog.
You can now view the ROI in any instance of vtkImageVisualizer, so the ROI can now be seen in the freehand calibration tasks.

Changed ImageFlip button be generic 2D actions button. Changed icon to suit new functionality. Added code to allow the definition of separators in the action menu for clarity.

git-svn-id: https://subversion.assembla.com/svn/plus/trunk@1925 c04e96f8-ac36-4ff2-8b34-4a5e34ef045f