re #551:
It seems that the function AcquisitionGrabberSapera::StopGrabbing() does not properly abort the transfer. 

Here's how it is done in AcquisitionGrabberSapera::StopGrabbing():

    BOOL result = saperaTransfer->Freeze();
    saperaTransfer->Wait(500);



And this is the code from AcquisitionGrabberSapera::Destroy():

     BOOL result = saperaTransfer->Freeze();
     if( saperaTransfer->Wait(500) != TRUE)
     {
         saperaTransfer->Abort();
      }

What is done in StopGrabbing() but not in Destroy(), is the cleanup of the injector:

    saperaProcessor->GetAcquisitionInjector()->CleanupInjection();

Demo_AcquisitionLibrary does not use stopGrabbing() at all, just Destroy().

So the solution could be not to use StopGrabbing(), but just Destroy().

With the changes in this commit, I am able to connect/disconnect repeatedly and close fCal without crash.




git-svn-id: https://subversion.assembla.com/svn/plus/trunk@1929 c04e96f8-ac36-4ff2-8b34-4a5e34ef045f