Re #553: added saving of the addition acquisition parameters.
Transducer name is saved as "BK-<model number><scan plane>". For 8848, scan plane can be either S or T. TransducerName attribute is added to RfProcessor, since it manages saving of that tag.

TransducerWidth initialization is added. BK OEM interface QUERY_SCANAREA returns Start/Stop Line X/Y in m. Start/StopLineX appear to be defined relative to the center of the active are of the transducer and are used to set transducer width (for linear probe only). Note that changing width on the console results in the change of Start/StopLineX (for sagittal array) and in the change of Start/StopLineAngle (for transverse). These changes can now be correctly captured via OEM interface. Tested passing to fCal configuration file that corresponds to S, while active surface was T, RfProcessor was correctly initialized.

TODO: Probably we need a bit smarter initialization of the reconstructed image size based on the dimensions of the scanned area and the width of the RF line. I will ask BK how we can query the RF line physical width, cannot find this in OEM specs. 


git-svn-id: https://subversion.assembla.com/svn/plus/trunk@1998 c04e96f8-ac36-4ff2-8b34-4a5e34ef045f