Fixed #569
Essentially, the reconstructed volume is created in the reference object coordinate frame. This was fine when the world coordinate frame was the same as the reference coordinate frame (as is the case in most config files). However, in the given config file the world coordinate frame is "Tracker" which breaks the visualization of the volume.

The solution was to create a displayable object attached to a coordinate frame to correct the missing transform. Now, simply create a displayable object of type polyData in the reference coordinate frame and assign it an ID.

Example:
<DisplayableObject Type="PolyData" ObjectCoordinateFrame="Reference" Id="Volume"/>

Then, in the config file section "VolumeReconstruction" you must tell it which ID is used to display the reconstructed volume.

Example:
<VolumeReconstruction
    ...
    VolumeDisplayableObjectId="Volume"
    ...
</>

git-svn-id: https://subversion.assembla.com/svn/plus/trunk@2007 c04e96f8-ac36-4ff2-8b34-4a5e34ef045f